Separating `Description` and `DescriptionItem` from `Metadata`.
