+++ src/modules/metadata/components/Metadata/index.js
@@ -1,107 +1,12 @@
 import './styles.css';
 import {
-  Anchor,
   Expandable,
   ExpandableButton,
-  ExpandableChildren,
-  Icon
+  ExpandableChildren
 } from '../../../reusable';
-import { BrowseLink } from '../../../browse';
+import Description from '../Description';
 import PropTypes from 'prop-types';
 import React from 'react';
-import { stringifySearch } from '../../../search';
-import { TrimString } from '../../../core';
-import { useSelector } from 'react-redux';
-
-const DescriptionItem = ({ browse, children, href, search }) => {
-  const { active: activeInstitution, defaultInstitution } = useSelector((state) => {
-    return state.institution;
-  });
-  const { active: activeDatastore, datastores } = useSelector((state) => {
-    return state.datastores;
-  });
-
-  const anchorAttributes = { href };
-
-  if (search) {
-    const { slug, uid } = datastores.find((datastore) => {
-      return datastore.uid === activeDatastore;
-    });
-    const { scope, type, value } = search;
-    anchorAttributes.to = `/${slug}?${stringifySearch({
-      filter: type === 'filtered' ? { [scope]: value } : {},
-      library: uid === 'mirlyn' ? (activeInstitution || defaultInstitution) : {},
-      query: type === 'fielded' ? `${scope}:${value}` : value
-    })}`;
-  }
-
-  const { text, type, value } = browse || {};
-
-  return (
-    <>
-      {(href || search) ? <Anchor {...anchorAttributes}>{children}</Anchor> : children}
-      {browse?.text && (
-        <>
-          <span className='text-grey font-small margin-x__2xs'>|</span>
-          <BrowseLink
-            className='text-grey font-small underline underline__hover-thick'
-            type={type}
-            value={value}
-          >
-            {text}
-          </BrowseLink>
-        </>
-      )}
-    </>
-  );
-};
-
-DescriptionItem.propTypes = {
-  browse: PropTypes.object,
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node
-  ]),
-  href: PropTypes.string,
-  search: PropTypes.object
-};
-
-const Description = ({ data, viewType }) => {
-  if (Array.isArray(data)) {
-    return (
-      <ol className='list__unstyled'>
-        {data.map((datum, index) => {
-          return (
-            <li key={index}>
-              {index > 0 && <Icon icon='navigate_next' className='text-grey__light' />}
-              <Description data={datum} />
-            </li>
-          );
-        })}
-      </ol>
-    );
-  }
-
-  const { icon, image, search: { scope } = {}, text } = data;
-
-  return (
-    <DescriptionItem {...data}>
-      <span style={{ display: image ? 'block' : 'initial' }}>
-        {icon && <Icon icon={icon} size={19} className='margin-right__2xs text-grey__light' />}
-        { viewType !== 'Full' && !data.search ? <TrimString {...{ string: text, trimLength: scope === 'author' ? 64 : 240 }} /> : text }
-      </span>
-      {image && <img src={image} alt='' className='padding-top__xs' />}
-    </DescriptionItem>
-  );
-};
-
-Description.propTypes = {
-  data: PropTypes.oneOfType([
-    PropTypes.array,
-    PropTypes.object
-  ]),
-  viewType: PropTypes.string
-};
 
 const Metadata = ({ data, viewType }) => {
   return (
@@ -130,7 +35,7 @@ const Metadata = ({ data, viewType }) => {
       })}
     </dl>
   );
-}
+};
 
 Metadata.propTypes = {
   data: PropTypes.array,
